+++ NEWS.md
@@ -5,9 +5,13 @@ Minimum LaBB-CAT version *20230224.1731*
 ## Enhancements
 
 - *getMatches* replaces *participant.ids* and *transcript.types* parameters with *participant.expression* and *transcript.expression* parameters, in line with the underlying API, and allowing more flexibility with queries.
+- *getMatches* parameter *aligned* is now deprecated, please set *anchor.confidence.min* = 50 instead
+- *getMatches* parameter *anchor.confidence.min* now filters out results, instead of just returning NA for offsets below the minumum confidence. To return all results and offsets, regardless of confidence, set *anchor.confidence.min* = 0.
+- *processWithPraat*'s *window.offset* parameter no longer has a default value, and includes more thorough documentation, to ensure callers are aware of the parameter's implications. 
 
 ## New functions:
 
+- *labbcatVersionInfo* - gets version information of all components of LaBB-CAT
 - Helper functions for formulating the new *participant.expression* and *transcript.expression* parameters of *getMatches*:
   + *expressionFromIds* : expression based on a list of IDs
   + *expressionFromAttributeValue* : expression based on a list of possible attribute values